changed back directory name to add more samples
