+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts-withCert/AppCreationScripts.md
