+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts-withCert/Cleanup.ps1
