+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts-withCert/Configure.ps1
