+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts-withCert/ConfigureVCService.ps1
