+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts/AppCreationScripts.md
