+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts/Cleanup.ps1
