+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts/Configure.ps1
