+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts/ConfigureVCService.ps1
