+++ 1-asp-net-core-api-idtokenhint/AppCreationScripts/createdApps.html
