+++ 1-asp-net-core-api-idtokenhint/AspNetCoreVerifiableCredentials.csproj
