+++ 1-asp-net-core-api-idtokenhint/Pages/Error.cshtml.cs
