+++ 1-asp-net-core-api-idtokenhint/Pages/Index.cshtml.cs
