+++ 1-asp-net-core-api-idtokenhint/Pages/Issuer.cshtml
