+++ 1-asp-net-core-api-idtokenhint/Pages/Shared/_ValidationScriptsPartial.cshtml
