+++ 1-asp-net-core-api-idtokenhint/Pages/Verifier.cshtml.cs
