+++ 1-asp-net-core-api-idtokenhint/Pages/_ViewStart.cshtml
