+++ 1-asp-net-core-api-idtokenhint/VerifierController.cs
@@ -1,19 +1,18 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.IO;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Options;
-using System.Security.Cryptography;
-using Newtonsoft.Json.Linq;
+using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Identity.Client;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+using System;
+using System.Diagnostics;
+using System.IO;
 using System.Net;
 using System.Net.Http;
 using System.Text;
-using Microsoft.Extensions.Caching.Memory;
-using System.Diagnostics;
+using System.Threading.Tasks;
 using System.Security.Cryptography.X509Certificates;
-using Microsoft.Identity.Client;
 using System.Net.Http.Headers;
+using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Logging;
 using Microsoft.Identity.Web;
 using Microsoft.Extensions.DependencyInjection;
@@ -21,16 +20,16 @@ using Microsoft.Extensions.DependencyInjection;
 namespace AspNetCoreVerifiableCredentials
 {
     [Route("api/[controller]/[action]")]
-    [ApiController]
-    public class IssuerController : ControllerBase
+    public class VerifierController : Controller
     {
-        const string ISSUANCEPAYLOAD = "issuance_request_config.json";
+        const string PRESENTATIONPAYLOAD = "presentation_request_config.json";
+        //        const string PRESENTATIONPAYLOAD = "presentation_request_config - TrueIdentitySample.json";
 
         protected readonly AppSettingsModel AppSettings;
         protected IMemoryCache _cache;
-        protected readonly ILogger<IssuerController> _log;
+        protected readonly ILogger<VerifierController> _log;
 
-        public IssuerController(IOptions<AppSettingsModel> appSettings, IMemoryCache memoryCache, ILogger<IssuerController> log)
+        public VerifierController(IOptions<AppSettingsModel> appSettings, IMemoryCache memoryCache, ILogger<VerifierController> log)
         {
             this.AppSettings = appSettings.Value;
             _cache = memoryCache;
@@ -38,78 +37,60 @@ namespace AspNetCoreVerifiableCredentials
         }
 
         /// <summary>
-        /// This method is called from the UI to initiate the issuance of the verifiable credential
+        /// This method is called from the UI to initiate the presentation of the verifiable credential
         /// </summary>
         /// <returns>JSON object with the address to the presentation request and optionally a QR code and a state value which can be used to check on the response status</returns>
-        [HttpGet("/api/issuer/issuance-request")]
-        public async Task<ActionResult> IssuanceRequest()
+        [HttpGet("/api/verifier/presentation-request")]
+        public async Task<ActionResult> PresentationRequest()
         {
             try
             {
+
+                string jsonString = null;
                 //they payload template is loaded from disk and modified in the code below to make it easier to get started
                 //and having all config in a central location appsettings.json. 
                 //if you want to manually change the payload in the json file make sure you comment out the code below which will modify it automatically
                 //
-                string jsonString = null;
-                string newpin = null;
-
-                string payloadpath = Path.Combine(Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location), ISSUANCEPAYLOAD);
+                string payloadpath = Path.Combine(Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location), PRESENTATIONPAYLOAD);
                 _log.LogTrace("IssuanceRequest file: {0}", payloadpath);
                 if (!System.IO.File.Exists(payloadpath))
                 {
                     _log.LogError("File not found: {0}", payloadpath);
-                    return BadRequest(new { error = "400", error_description = ISSUANCEPAYLOAD + " not found" });
+                    return BadRequest(new { error = "400", error_description = PRESENTATIONPAYLOAD + " not found" });
                 }
                 jsonString = System.IO.File.ReadAllText(payloadpath);
                 if (string.IsNullOrEmpty(jsonString))
                 {
                     _log.LogError("Error reading file: {0}", payloadpath);
-                    return BadRequest(new { error = "400", error_description = ISSUANCEPAYLOAD + " error reading file" });
-                }
-
-                //check if pin is required, if found make sure we set a new random pin
-                //pincode is only used when the payload contains claim value pairs which results in an IDTokenhint
-                JObject payload = JObject.Parse(jsonString);
-                if (payload["issuance"]["pin"] != null)
-                {
-                    if (IsMobile())
-                    {
-                        _log.LogTrace("pin element found in JSON payload, but on mobile so remove pin since we will be using deeplinking");
-                        //consider providing the PIN through other means to your user instead of removing it.
-                        payload["issuance"]["pin"].Parent.Remove();
-
-                    }
-                    else
-                    {
-                        _log.LogTrace("pin element found in JSON payload, modifying to a random number of the specific length");
-                        var length = (int)payload["issuance"]["pin"]["length"];
-                        var pinMaxValue = (int)Math.Pow(10, length) - 1;
-                        var randomNumber = RandomNumberGenerator.GetInt32(1, pinMaxValue);
-                        newpin = string.Format("{0:D" + length.ToString() + "}", randomNumber);
-                        payload["issuance"]["pin"]["value"] = newpin;
-                    }
-
+                    return BadRequest(new { error = "400", error_description = PRESENTATIONPAYLOAD + " error reading file" });
                 }
 
                 string state = Guid.NewGuid().ToString();
 
                 //modify payload with new state, the state is used to be able to update the UI when callbacks are received from the VC Service
+                JObject payload = JObject.Parse(jsonString);
                 if (payload["callback"]["state"] != null)
                 {
                     payload["callback"]["state"] = state;
                 }
 
-                //get the IssuerDID from the appsettings
+                //get the VerifierDID from the appsettings
                 if (payload["authority"] != null)
                 {
-                    payload["authority"] = AppSettings.IssuerAuthority;
+                    payload["authority"] = AppSettings.VerifierAuthority;
                 }
 
-                //modify the callback method to make it easier to debug 
-                //with tools like ngrok since the URI changes all the time
-                //this way you don't need to modify the callback URL in the payload every time
-                //ngrok changes the URI
+                //copy the issuerDID from the settings and fill in the trustedIssuer part of the payload
+                //this means only that issuer should be trusted for the requested credentialtype
+                //this value is an array in the payload, you can trust multiple issuers for the same credentialtype
+                //very common to accept the test VCs and the Production VCs coming from different verifiable credential services
+                if (payload["presentation"]["requestedCredentials"][0]["acceptedIssuers"][0] != null)
+                {
+                    payload["presentation"]["requestedCredentials"][0]["acceptedIssuers"][0] = AppSettings.IssuerAuthority;
+                }
 
+                //modify the callback method to make it easier to debug with tools like ngrok since the URI changes all the time
+                //this way you don't need to modify the callback URL in the payload every time ngrok changes the URI
                 if (payload["callback"]["url"] != null)
                 {
                     //localhost hostname can't work for callbacks so we won't overwrite it.
@@ -118,29 +99,16 @@ namespace AspNetCoreVerifiableCredentials
                     string host = GetRequestHostName();
                     if (!host.Contains("//localhost"))
                     {
-                        payload["callback"]["url"] = String.Format("{0}:/api/issuer/issuanceCallback", host);
+                        payload["callback"]["url"] = String.Format("{0}:/api/verifier/presentationCallback", host);
                     }
                 }
 
-                //get the manifest from the appsettings, this is the URL to the credential created in the azure portal. 
-                //the display and rules file to create the credential can be dound in the credentialfiles directory
-                //make sure the credentialtype in the issuance payload matches with the rules file
-                //for this sample it should be VerifiedCredentialExpert
-                if (payload["issuance"]["manifest"] != null)
-                {
-                    payload["issuance"]["manifest"] = AppSettings.CredentialManifest;
-                }
-
-                //here you could change the payload manifest and change the firstname and lastname
-                payload["issuance"]["claims"]["given_name"] = "Megan";
-                payload["issuance"]["claims"]["family_name"] = "Bowen";
-
                 jsonString = JsonConvert.SerializeObject(payload);
+
 
                 //CALL REST API WITH PAYLOAD
                 HttpStatusCode statusCode = HttpStatusCode.OK;
                 string response = null;
-
                 try
                 {
                     //The VC Request API is an authenticated API. We need to clientid and secret (or certificate) to create an access token which 
@@ -152,21 +120,19 @@ namespace AspNetCoreVerifiableCredentials
                         return BadRequest(new { error = accessToken.error, error_description = accessToken.error_description });
                     }
 
-
                     HttpClient client = new HttpClient();
                     var defaultRequestHeaders = client.DefaultRequestHeaders;
                     defaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", accessToken.token);
 
                     HttpResponseMessage res = await client.PostAsync(AppSettings.ApiEndpoint, new StringContent(jsonString, Encoding.UTF8, "application/json"));
                     response = await res.Content.ReadAsStringAsync();
+                    _log.LogTrace("succesfully called Request API");
                     client.Dispose();
                     statusCode = res.StatusCode;
 
                     if (statusCode == HttpStatusCode.Created)
                     {
-                        _log.LogTrace("succesfully called Request API");
                         JObject requestConfig = JObject.Parse(response);
-                        if (newpin != null) { requestConfig["pin"] = newpin; }
                         requestConfig.Add(new JProperty("id", state));
                         jsonString = JsonConvert.SerializeObject(requestConfig);
 
@@ -180,6 +146,10 @@ namespace AspNetCoreVerifiableCredentials
                         };
                         _cache.Set(state, JsonConvert.SerializeObject(cacheData));
 
+                        //the response from the VC Request API call is returned to the caller (the UI). It contains the URI to the request which Authenticator can download after
+                        //it has scanned the QR code. If the payload requested the VC Request service to create the QR code that is returned as well
+                        //the javascript in the UI will use that QR code to display it on the screen to the user.
+
                         return new ContentResult { ContentType = "application/json", Content = jsonString };
                     }
                     else
@@ -187,7 +157,6 @@ namespace AspNetCoreVerifiableCredentials
                         _log.LogError("Unsuccesfully called Request API");
                         return BadRequest(new { error = "400", error_description = "Something went wrong calling the API: " + response });
                     }
-
                 }
                 catch (Exception ex)
                 {
@@ -201,61 +170,52 @@ namespace AspNetCoreVerifiableCredentials
         }
 
         /// <summary>
-        /// This method is called by the VC Request API when the user scans a QR code and accepts the issued Verifiable Credential
+        /// This method is called by the VC Request API when the user scans a QR code and presents a Verifiable Credential to the service
         /// </summary>
         /// <returns></returns>
         [HttpPost]
-        public async Task<ActionResult> IssuanceCallback()
+        public async Task<ActionResult> PresentationCallback()
         {
             try
             {
                 string content = await new System.IO.StreamReader(this.Request.Body).ReadToEndAsync();
-                _log.LogTrace("callback!: " + content);
-                JObject issuanceResponse = JObject.Parse(content);
-                var state = issuanceResponse["state"].ToString();
+                Debug.WriteLine("callback!: " + content);
+                JObject presentationResponse = JObject.Parse(content);
+                var state = presentationResponse["state"].ToString();
 
                 //there are 2 different callbacks. 1 if the QR code is scanned (or deeplink has been followed)
                 //Scanning the QR code makes Authenticator download the specific request from the server
                 //the request will be deleted from the server immediately.
                 //That's why it is so important to capture this callback and relay this to the UI so the UI can hide
                 //the QR code to prevent the user from scanning it twice (resulting in an error since the request is already deleted)
-                if (issuanceResponse["code"].ToString() == "request_retrieved")
+                if (presentationResponse["code"].ToString() == "request_retrieved")
                 {
                     var cacheData = new
                     {
                         status = "request_retrieved",
-                        message = "QR Code is scanned. Waiting for issuance...",
+                        message = "QR Code is scanned. Waiting for validation...",
                     };
                     _cache.Set(state, JsonConvert.SerializeObject(cacheData));
                 }
 
-                //
-                //This callback is called when issuance is completed.
-                //
-                if (issuanceResponse["code"].ToString() == "issuance_successful")
+                // the 2nd callback is the result with the verified credential being verified.
+                // typically here is where the business logic is written to determine what to do with the result
+                // the response in this callback contains the claims from the Verifiable Credential(s) being presented by the user
+                // In this case the result is put in the in memory cache which is used by the UI when polling for the state so the UI can be updated.
+                if (presentationResponse["code"].ToString() == "presentation_verified")
                 {
                     var cacheData = new
                     {
-                        status = "issuance_successful",
-                        message = "Credential successfully issued",
-                    };
-                    _cache.Set(state, JsonConvert.SerializeObject(cacheData));
-                }
-                //
-                //We capture if something goes wrong during issuance. See documentation with the different error codes
-                //
-                if (issuanceResponse["code"].ToString() == "issuance_error")
-                {
-                    var cacheData = new
-                    {
-                        status = "issuance_error",
-                        payload = issuanceResponse["error"]["code"].ToString(),
-                        //at the moment there isn't a specific error for incorrect entry of a pincode.
-                        //So assume this error happens when the users entered the incorrect pincode and ask to try again.
-                        message = issuanceResponse["error"]["message"].ToString()
+                        status = "presentation_verified",
+                        message = "Presentation verified",
+                        payload = presentationResponse["issuers"].ToString(),
+                        subject = presentationResponse["subject"].ToString(),
+                        firstName = presentationResponse["issuers"][0]["claims"]["firstName"].ToString(),
+                        lastName = presentationResponse["issuers"][0]["claims"]["lastName"].ToString()
 
                     };
                     _cache.Set(state, JsonConvert.SerializeObject(cacheData));
+
                 }
 
                 return new OkResult();
@@ -265,15 +225,15 @@ namespace AspNetCoreVerifiableCredentials
                 return BadRequest(new { error = "400", error_description = ex.Message });
             }
         }
-
         //
         //this function is called from the UI polling for a response from the AAD VC Service.
-        //when a callback is recieved at the issuanceCallback service the session will be updated
-        //this method will respond with the status so the UI can reflect if the QR code was scanned and with the result of the issuance process
+        //when a callback is recieved at the presentationCallback service the session will be updated
+        //this method will respond with the status so the UI can reflect if the QR code was scanned and with the result of the presentation
         //
-        [HttpGet("/api/issuer/issuance-response")]
-        public ActionResult IssuanceResponse()
+        [HttpGet("/api/verifier/presentation-response")]
+        public ActionResult PresentationResponse()
         {
+
             try
             {
                 //the id is the state value initially created when the issuanc request was requested from the request API
@@ -298,12 +258,13 @@ namespace AspNetCoreVerifiableCredentials
             {
                 return BadRequest(new { error = "400", error_description = ex.Message });
             }
+
+
         }
 
         //some helper functions
         protected async Task<(string token, string error, string error_description)> GetAccessToken()
         {
-
             // You can run this sample using ClientSecret or Certificate. The code will differ only when instantiating the IConfidentialClientApplication
             bool isUsingClientSecret = AppSettings.AppUsesClientSecret(AppSettings);
 
@@ -371,16 +332,6 @@ namespace AspNetCoreVerifiableCredentials
                 hostname = string.Format("{0}://{1}", scheme, originalHost);
             else hostname = string.Format("{0}://{1}", scheme, this.Request.Host);
             return hostname;
-        }
-
-        protected bool IsMobile()
-        {
-            string userAgent = this.Request.Headers["User-Agent"];
-
-            if (userAgent.Contains("Android") || userAgent.Contains("iPhone"))
-                return true;
-            else
-                return false;
         }
     }
 }
