+++ 1-asp-net-core-api-idtokenhint/wwwroot/authenticator-icon.png
