+++ 1-asp-net-core-api-idtokenhint/wwwroot/favicon.png
